+++ core/src/main/java/uk/ac/bbsrc/tgac/miso/core/store/LibraryDesignDao.java
@@ -6,15 +6,12 @@ import java.util.List;
 import uk.ac.bbsrc.tgac.miso.core.data.LibraryDesign;
 import uk.ac.bbsrc.tgac.miso.core.data.SampleClass;
 
-public interface LibraryDesignDao {
-
-  /**
-   * @return a list of all LibraryDesigns
-   */
-  List<LibraryDesign> getLibraryDesigns() throws IOException;
+public interface LibraryDesignDao extends SaveDao<LibraryDesign> {
   
-  List<LibraryDesign> getLibraryDesignByClass(SampleClass sampleClass) throws IOException;
+  public LibraryDesign getByNameAndSampleClass(String name, SampleClass sampleClass) throws IOException;
 
-  LibraryDesign getLibraryDesign(Long id) throws IOException;
+  public List<LibraryDesign> listByClass(SampleClass sampleClass) throws IOException;
+
+  public long getUsage(LibraryDesign design) throws IOException;
 
 }