+++ sqlstore/src/test/java/uk/ac/bbsrc/tgac/miso/persistence/impl/HibernateArrayModelDaoTest.java
@@ -5,25 +5,20 @@ import static org.junit.Assert.*;
 import java.io.IOException;
 import java.util.List;
 
-import org.hibernate.SessionFactory;
 import org.junit.Before;
 import org.junit.Test;
-import org.springframework.beans.factory.annotation.Autowired;
 
 import uk.ac.bbsrc.tgac.miso.AbstractDAOTest;
 import uk.ac.bbsrc.tgac.miso.core.data.ArrayModel;
 
 public class HibernateArrayModelDaoTest extends AbstractDAOTest {
 
-  @Autowired
-  private SessionFactory sessionFactory;
-
   private HibernateArrayModelDao sut;
 
   @Before
   public void setup() {
     sut = new HibernateArrayModelDao();
-    sut.setSessionFactory(sessionFactory);
+    sut.setSessionFactory(getSessionFactory());
   }
 
   @Test
@@ -60,7 +55,7 @@ public class HibernateArrayModelDaoTest extends AbstractDAOTest {
 
     clearSession();
 
-    ArrayModel saved = (ArrayModel) sessionFactory.getCurrentSession().get(ArrayModel.class, savedId);
+    ArrayModel saved = (ArrayModel) getSessionFactory().getCurrentSession().get(ArrayModel.class, savedId);
     assertEquals(alias, saved.getAlias());
   }
 
@@ -68,27 +63,22 @@ public class HibernateArrayModelDaoTest extends AbstractDAOTest {
   public void testUpdate() throws IOException {
     long id = 1L;
     String alias = "New Name";
-    ArrayModel arrayModel = (ArrayModel) sessionFactory.getCurrentSession().get(ArrayModel.class, id);
+    ArrayModel arrayModel = (ArrayModel) getSessionFactory().getCurrentSession().get(ArrayModel.class, id);
     assertNotEquals(alias, arrayModel.getAlias());
     arrayModel.setAlias(alias);
     sut.update(arrayModel);
 
     clearSession();
 
-    ArrayModel saved = (ArrayModel) sessionFactory.getCurrentSession().get(ArrayModel.class, id);
+    ArrayModel saved = (ArrayModel) getSessionFactory().getCurrentSession().get(ArrayModel.class, id);
     assertEquals(alias, saved.getAlias());
   }
 
   @Test
   public void testGetUsage() throws IOException {
-    ArrayModel arrayModel = (ArrayModel) sessionFactory.getCurrentSession().get(ArrayModel.class, 1L);
+    ArrayModel arrayModel = (ArrayModel) getSessionFactory().getCurrentSession().get(ArrayModel.class, 1L);
     assertEquals("Test BeadChip", arrayModel.getAlias());
     assertEquals(1L, sut.getUsage(arrayModel));
-  }
-
-  private void clearSession() {
-    sessionFactory.getCurrentSession().flush();
-    sessionFactory.getCurrentSession().clear();
   }
 
 }