+++ sqlstore/src/test/java/uk/ac/bbsrc/tgac/miso/persistence/impl/HibernatePartitionQcTypeDaoTest.java
@@ -5,25 +5,20 @@ import static org.junit.Assert.*;
 import java.io.IOException;
 import java.util.List;
 
-import org.hibernate.SessionFactory;
 import org.junit.Before;
 import org.junit.Test;
-import org.springframework.beans.factory.annotation.Autowired;
 
 import uk.ac.bbsrc.tgac.miso.AbstractDAOTest;
 import uk.ac.bbsrc.tgac.miso.core.data.PartitionQCType;
 
 public class HibernatePartitionQcTypeDaoTest extends AbstractDAOTest {
 
-  @Autowired
-  private SessionFactory sessionFactory;
-
   private HibernatePartitionQcTypeDao sut;
 
   @Before
   public void setup() {
     sut = new HibernatePartitionQcTypeDao();
-    sut.setSessionFactory(sessionFactory);
+    sut.setSessionFactory(getSessionFactory());
   }
 
   @Test
@@ -58,7 +53,7 @@ public class HibernatePartitionQcTypeDaoTest extends AbstractDAOTest {
 
     clearSession();
 
-    PartitionQCType saved = (PartitionQCType) sessionFactory.getCurrentSession().get(PartitionQCType.class, savedId);
+    PartitionQCType saved = (PartitionQCType) getSessionFactory().getCurrentSession().get(PartitionQCType.class, savedId);
     assertEquals(desc, saved.getDescription());
   }
 
@@ -66,27 +61,22 @@ public class HibernatePartitionQcTypeDaoTest extends AbstractDAOTest {
   public void testUpdate() throws IOException {
     long typeId = 1L;
     String newDesc = "New Desc";
-    PartitionQCType st = (PartitionQCType) sessionFactory.getCurrentSession().get(PartitionQCType.class, typeId);
+    PartitionQCType st = (PartitionQCType) getSessionFactory().getCurrentSession().get(PartitionQCType.class, typeId);
     assertNotEquals(newDesc, st.getDescription());
     st.setDescription(newDesc);
     assertEquals(typeId, sut.update(st));
 
     clearSession();
 
-    PartitionQCType saved = (PartitionQCType) sessionFactory.getCurrentSession().get(PartitionQCType.class, typeId);
+    PartitionQCType saved = (PartitionQCType) getSessionFactory().getCurrentSession().get(PartitionQCType.class, typeId);
     assertEquals(newDesc, saved.getDescription());
   }
 
   @Test
   public void testGetUsage() throws IOException {
-    PartitionQCType ok = (PartitionQCType) sessionFactory.getCurrentSession().get(PartitionQCType.class, 1L);
+    PartitionQCType ok = (PartitionQCType) getSessionFactory().getCurrentSession().get(PartitionQCType.class, 1L);
     assertEquals("OK", ok.getDescription());
     assertEquals(1L, sut.getUsage(ok));
-  }
-
-  private void clearSession() {
-    sessionFactory.getCurrentSession().flush();
-    sessionFactory.getCurrentSession().clear();
   }
 
 }