+++ sqlstore/src/test/java/uk/ac/bbsrc/tgac/miso/persistence/impl/HibernateSampleTypeDaoTest.java
@@ -5,25 +5,20 @@ import static org.junit.Assert.*;
 import java.io.IOException;
 import java.util.List;
 
-import org.hibernate.SessionFactory;
 import org.junit.Before;
 import org.junit.Test;
-import org.springframework.beans.factory.annotation.Autowired;
 
 import uk.ac.bbsrc.tgac.miso.AbstractDAOTest;
 import uk.ac.bbsrc.tgac.miso.core.data.SampleType;
 
 public class HibernateSampleTypeDaoTest extends AbstractDAOTest {
 
-  @Autowired
-  private SessionFactory sessionFactory;
-
   private HibernateSampleTypeDao sut;
 
   @Before
   public void setup() {
     sut = new HibernateSampleTypeDao();
-    sut.setSessionFactory(sessionFactory);
+    sut.setSessionFactory(getSessionFactory());
   }
 
   @Test
@@ -57,7 +52,7 @@ public class HibernateSampleTypeDaoTest extends AbstractDAOTest {
 
     clearSession();
 
-    SampleType saved = (SampleType) sessionFactory.getCurrentSession().get(SampleType.class, savedId);
+    SampleType saved = (SampleType) getSessionFactory().getCurrentSession().get(SampleType.class, savedId);
     assertEquals(name, saved.getName());
   }
 
@@ -65,31 +60,26 @@ public class HibernateSampleTypeDaoTest extends AbstractDAOTest {
   public void testUpdate() throws IOException {
     long typeId = 1L;
     String newName = "New Name";
-    SampleType st = (SampleType) sessionFactory.getCurrentSession().get(SampleType.class, typeId);
+    SampleType st = (SampleType) getSessionFactory().getCurrentSession().get(SampleType.class, typeId);
     assertNotEquals(newName, st.getName());
     st.setName(newName);
     assertEquals(typeId, sut.update(st));
 
     clearSession();
 
-    SampleType saved = (SampleType) sessionFactory.getCurrentSession().get(SampleType.class, typeId);
+    SampleType saved = (SampleType) getSessionFactory().getCurrentSession().get(SampleType.class, typeId);
     assertEquals(newName, saved.getName());
   }
 
   @Test
   public void testGetUsage() throws IOException {
-    SampleType genomic = (SampleType) sessionFactory.getCurrentSession().get(SampleType.class, 1L);
+    SampleType genomic = (SampleType) getSessionFactory().getCurrentSession().get(SampleType.class, 1L);
     assertEquals("GENOMIC", genomic.getName());
     assertTrue(sut.getUsage(genomic) > 10L);
 
-    SampleType synthetic = (SampleType) sessionFactory.getCurrentSession().get(SampleType.class, 3L);
+    SampleType synthetic = (SampleType) getSessionFactory().getCurrentSession().get(SampleType.class, 3L);
     assertEquals("SYNTHETIC", synthetic.getName());
     assertEquals(0L, sut.getUsage(synthetic));
-  }
-
-  private void clearSession() {
-    sessionFactory.getCurrentSession().flush();
-    sessionFactory.getCurrentSession().clear();
   }
 
 }