+++ sqlstore/src/test/java/uk/ac/bbsrc/tgac/miso/persistence/impl/HibernateTissueTypeDaoTest.java
@@ -5,10 +5,8 @@ import static org.junit.Assert.*;
 import java.io.IOException;
 import java.util.List;
 
-import org.hibernate.SessionFactory;
 import org.junit.Before;
 import org.junit.Test;
-import org.springframework.beans.factory.annotation.Autowired;
 
 import com.eaglegenomics.simlims.core.User;
 
@@ -19,15 +17,12 @@ import uk.ac.bbsrc.tgac.miso.core.data.impl.UserImpl;
 
 public class HibernateTissueTypeDaoTest extends AbstractDAOTest {
 
-  @Autowired
-  private SessionFactory sessionFactory;
-
   private HibernateTissueTypeDao sut;
 
   @Before
   public void setup() {
     sut = new HibernateTissueTypeDao();
-    sut.setSessionFactory(sessionFactory);
+    sut.setSessionFactory(getSessionFactory());
   }
 
   @Test
@@ -59,13 +54,13 @@ public class HibernateTissueTypeDaoTest extends AbstractDAOTest {
     TissueType tt = new TissueTypeImpl();
     tt.setAlias(alias);
     tt.setDescription("descriptive stuff");
-    User user = (User) sessionFactory.getCurrentSession().get(UserImpl.class, 1L);
+    User user = (User) getSessionFactory().getCurrentSession().get(UserImpl.class, 1L);
     tt.setChangeDetails(user);
     long savedId = sut.create(tt);
 
     clearSession();
 
-    TissueType saved = (TissueType) sessionFactory.getCurrentSession().get(TissueTypeImpl.class, savedId);
+    TissueType saved = (TissueType) getSessionFactory().getCurrentSession().get(TissueTypeImpl.class, savedId);
     assertEquals(alias, saved.getAlias());
   }
 
@@ -73,27 +68,22 @@ public class HibernateTissueTypeDaoTest extends AbstractDAOTest {
   public void testUpdate() throws IOException {
     long id = 1L;
     String alias = "New Name";
-    TissueType tt = (TissueType) sessionFactory.getCurrentSession().get(TissueTypeImpl.class, id);
+    TissueType tt = (TissueType) getSessionFactory().getCurrentSession().get(TissueTypeImpl.class, id);
     assertNotEquals(alias, tt.getAlias());
     tt.setAlias(alias);
     sut.update(tt);
 
     clearSession();
 
-    TissueType saved = (TissueType) sessionFactory.getCurrentSession().get(TissueTypeImpl.class, id);
+    TissueType saved = (TissueType) getSessionFactory().getCurrentSession().get(TissueTypeImpl.class, id);
     assertEquals(alias, saved.getAlias());
   }
 
   @Test
   public void testGetUsage() throws IOException {
-    TissueType tt = (TissueType) sessionFactory.getCurrentSession().get(TissueTypeImpl.class, 1L);
+    TissueType tt = (TissueType) getSessionFactory().getCurrentSession().get(TissueTypeImpl.class, 1L);
     assertEquals("Test Type", tt.getAlias());
     assertEquals(2L, sut.getUsage(tt));
-  }
-
-  private void clearSession() {
-    sessionFactory.getCurrentSession().flush();
-    sessionFactory.getCurrentSession().clear();
   }
 
 }